requirements.txt
torch>=2.0.0
transformers>=4.40.0
langchain>=0.1.0
pypdf>=4.2.0
sentence-transformers>=2.6.0
accelerate>=0.29.0
peft>=0.10.0
trl>=0.8.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
datasets>=2.18.0
gradio>=4.25.0